feat(UIKIT-1163,PageHeader): Правки по PR
